+++ test/dummy/app/controllers/pages_controller.rb
@@ -1,11 +1,7 @@
 class PagesController < ApplicationController
   before_action :load_workspace
   before_action :load_root_recording
-  before_action :authorize_view_root!, only: %i[index]
-  before_action :authorize_edit_root!, only: %i[new create]
   before_action :load_recording, only: %i[show edit update destroy restore]
-  before_action :authorize_view_recording!, only: %i[show]
-  before_action :authorize_edit_recording!, only: %i[edit update destroy restore]
 
   def index
     scope = @root_recording.recordings_of(RecordingStudioPage)
@@ -57,7 +53,7 @@ class PagesController < ApplicationController
   private
 
   def load_workspace
-    @workspace = Workspace.first_or_create!(name: "Studio Workspace")
+    @workspace = Workspace.order(:created_at).first_or_create!(name: "Studio Workspace")
   end
 
   def load_root_recording
@@ -65,8 +61,6 @@ class PagesController < ApplicationController
       recordable: @workspace,
       parent_recording_id: nil
     )
-
-    ensure_root_access!(@workspace)
   end
 
   def load_recording
@@ -75,21 +69,5 @@ class PagesController < ApplicationController
 
   def page_params
     params.require(:page).permit(:title, :summary)
-  end
-
-  def authorize_view_root!
-    require_root_access!(@root_recording, minimum_role: :view)
-  end
-
-  def authorize_edit_root!
-    require_root_access!(@root_recording, minimum_role: :edit)
-  end
-
-  def authorize_view_recording!
-    require_recording_access!(@recording, minimum_role: :view)
-  end
-
-  def authorize_edit_recording!
-    require_recording_access!(@recording, minimum_role: :edit)
   end
 end