+++ test/models/event_test.rb
@@ -9,7 +9,6 @@ class EventTest < ActiveSupport::TestCase
     RecordingStudio::DelegatedTypeRegistrar.apply!
 
     RecordingStudio::Event.delete_all
-    RecordingStudio::DeviceSession.delete_all
     RecordingStudio::Recording.delete_all
     RecordingStudioPage.delete_all
     Workspace.delete_all