+++ test/recording_studio_test.rb
@@ -1,19 +1,13 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "stringio"
 
 class RecordingStudioTest < Minitest::Test
   def setup
-    @original_features = RecordingStudio.features.to_h
     @original_registered_capabilities = RecordingStudio.registered_capabilities.dup
-    RecordingStudio.reset_runtime_warnings!
   end
 
   def teardown
-    @original_features.each do |feature_name, value|
-      RecordingStudio.features.public_send("#{feature_name}=", value)
-    end
     RecordingStudio.instance_variable_set(:@registered_capabilities, @original_registered_capabilities)
   end
 
@@ -23,36 +17,6 @@ class RecordingStudioTest < Minitest::Test
 
   def test_engine_exists
     assert_kind_of Class, ::RecordingStudio::Engine
-  end
-
-  def test_feature_toggles_default_to_true
-    assert RecordingStudio.features.device_sessions?
-  end
-
-  def test_feature_toggles_can_be_updated
-    RecordingStudio.configure do |config|
-      config.features.device_sessions = false
-    end
-
-    assert_not RecordingStudio.features.device_sessions?
-  end
-
-  def test_feature_toggles_cast_common_string_values
-    RecordingStudio.configure do |config|
-      config.features.device_sessions = "0"
-    end
-
-    assert_not RecordingStudio.features.device_sessions?
-  end
-
-  def test_warn_legacy_feature_use_emits_once_per_feature
-    warnings = capture_logger_warnings do
-      RecordingStudio.warn_legacy_feature_use!(:device_sessions, used_by: "test")
-      RecordingStudio.warn_legacy_feature_use!(:device_sessions, used_by: "test")
-    end
-
-    assert_includes warnings, "config.features.device_sessions = false"
-    assert_equal 1, warnings.scan("Legacy built-in 'device_sessions' feature").size
   end
 
   def test_register_capability_applies_capability_without_manual_apply
@@ -72,31 +36,7 @@ class RecordingStudioTest < Minitest::Test
     assert_equal capability_module, RecordingStudio.registered_capabilities.fetch(:auto_apply_probe).fetch(:mod)
   end
 
-  def test_capability_registration_can_override_legacy_gate_for_addon_replacements
-    RecordingStudio.features.device_sessions = false
-
-    capability_module = Module.new do
-      def addon_device_session_probe_value
-        :ok
-      end
-    end
-
-    RecordingStudio.register_capability(:device_sessions, capability_module)
-
-    assert_includes RecordingStudio::Recording.included_modules, capability_module
-    assert RecordingStudio::Recording.new.respond_to?(:addon_device_session_probe_value)
-  end
-
-  private
-
-  def capture_logger_warnings
-    original_logger = Rails.logger
-    io = StringIO.new
-    Rails.logger = ActiveSupport::Logger.new(io)
-    Rails.logger.level = Logger::WARN
-    yield
-    io.string
-  ensure
-    Rails.logger = original_logger
+  def test_configuration_to_h_excludes_removed_feature_flags
+    refute_includes RecordingStudio.configuration.to_h.keys, :features
   end
 end